requirements.txt
python-telegram-bot==20.3
beautifulsoup4==4.13.0
requests==2.31.0
nltk==3.8.1
